# docs/swagger.yaml
definitions:
...
  wn_internal_domain_dto.DragNoteRequest:
    properties:
      noteId:
        type: string
      toLayoutId:
        type: string
    type: object
...
  wn_internal_domain_dto.LinkBetweenNotesRequest:
    properties:
      firstNoteId:
        type: string
      layoutId:
        type: string
      secondNoteId:
        type: string
    required:
    - firstNoteId
    - layoutId
    - secondNoteId
    type: object
...
  wn_internal_domain_dto.Note:
    properties:
      draft:
        type: string
      haveAccess:
        items:
          type: string
        type: array
      id:
        type: string
      layoutId:
        type: string
      linkedWithIn:
        items:
          type: string
        type: array
      linkedWithOut:
        items:
          type: string
        type: array
      ownerId:
        type: string
      payload:
        type: string
      position:
        $ref: '#/definitions/wn_internal_domain_dto.Position'
      title:
        type: string
    type: object
  wn_internal_domain_dto.NoteId:
    properties:
      noteId:
        type: string
    required:
    - noteId
    type: object
  wn_internal_domain_dto.NoteRequest:
    properties:
      layoutId:
        type: string
      payload:
        type: string
      title:
        type: string
    type: object
  wn_internal_domain_dto.NoteWithIdRequest:
    properties:
      noteId:
        type: string
      payload:
        type: string
      title:
        type: string
    required:
    - noteId
    type: object
...
  wn_internal_domain_dto.Position:
    properties:
      xPos:
        type: number
      yPos:
        type: number
    type: object
...
  wn_internal_domain_dto.UpdateNotePositionRequest:
    properties:
      layoutId:
        type: string
      noteId:
        type: string
      xPos:
        type: number
      yPos:
        type: number
    required:
    - layoutId
    - noteId
    type: object
...
  wn_pkg_response.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/wn_pkg_response.meta'
      pagination:
        $ref: '#/definitions/wn_pkg_response.pagination'
    type: object
  wn_pkg_response.meta:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
      requestId:
        type: string
    type: object
  wn_pkg_response.pagination:
    properties:
      page:
        type: integer
      pages:
        type: integer
      perPage:
        type: integer
    type: object
info:
  contact: {}
  description: This is walrus notes api service.
  title: WALRUS NOTES API
  version: "1.0"
paths:
...
  /wn/api/v1/notes/create:
    post:
      description: Создать заметку
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.NoteRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/wn_internal_domain_dto.NoteId'
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: not_unique, permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: create_note
      tags:
      - notes
  /wn/api/v1/notes/delete:
    post:
      description: Удалить заметку
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.NoteId'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: note_not_found, permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: delete_note
      tags:
      - notes
  /wn/api/v1/notes/drag:
    post:
      description: Переместить заметку между лейаутами
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.DragNoteRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "401":
          description: 'possible codes: invalid_token, invalid_authorization_header'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: permissions_not_enough, record_not_found'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: drag_note
      tags:
      - notes
  /wn/api/v1/notes/layout:
    get:
      description: Получить все заметки из layout-а
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: layoutId
        in: query
        name: layoutId
        required: true
        type: string
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/wn_internal_domain_dto.Note'
                  type: array
              type: object
        "400":
          description: 'possible codes: bind_query, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: note_not_found, permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: get_notes_from_layout
      tags:
      - notes
  /wn/api/v1/notes/layout/graph/note:
    post:
      description: Обновить позицию заметки в графе
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.UpdateNotePositionRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id, permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: update_note_position
      tags:
      - graph
  /wn/api/v1/notes/layout/graph/posed:
    get:
      description: Получить заметки, которые имеют позиции в графе
      parameters:
      - description: layoutId
        in: query
        name: layoutId
        required: true
        type: string
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/wn_internal_domain_dto.Note'
                  type: array
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id, permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: get_posed_notes
      tags:
      - graph
  /wn/api/v1/notes/layout/graph/unposed:
    get:
      description: Получить заметки, которые не имеют позиции в графе
      parameters:
      - description: layoutId
        in: query
        name: layoutId
        required: true
        type: string
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/wn_internal_domain_dto.Note'
                  type: array
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id, permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: get_unposed_notes
      tags:
      - graph
  /wn/api/v1/notes/layout/links/create:
    post:
      description: Связать заметки
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.LinkBetweenNotesRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id, permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: create_link_between_notes
      tags:
      - links
  /wn/api/v1/notes/layout/links/delete:
    post:
      description: Обновить позицию заметки в графе
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.LinkBetweenNotesRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id, permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: delete_link_between_notes
      tags:
      - links
  /wn/api/v1/notes/search:
    post:
      description: Найти заметку
      parameters:
      - description: search
        in: query
        name: search
        required: true
        type: string
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/wn_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/wn_internal_domain_dto.Note'
                  type: array
              type: object
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: search_notes
      tags:
      - notes
  /wn/api/v1/notes/update:
    post:
      description: Обновить заметку
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/wn_internal_domain_dto.NoteWithIdRequest'
      - description: Request id identity
        in: header
        name: X-Request-Id
        required: true
        type: string
      - description: auth token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "400":
          description: 'possible codes: bind_body, invalid_X-Request-Id'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
        "422":
          description: 'possible codes: note_not_found, permissions_not_enough'
          schema:
            $ref: '#/definitions/wn_pkg_response.Response'
      summary: update_note
      tags:
      - notes
...
swagger: "2.0"